+++ ejb/src/main/resources/database/V1_DDL_Tenant.sql
@@ -5,14 +5,14 @@ CREATE SCHEMA IF NOT EXISTS tenant_clinica1
 
 GRANT ALL PRIVILEGES ON SCHEMA tenant_clinica1 TO clinica1_user;
 GRANT ALL PRIVILEGES ON SCHEMA tenant_clinica1 TO clinica1_user;
-GRANT ALL PRIVILEGES ON ALL TABLES IN SCHEMA tenant_clinica1         TO clinica1_user;
-GRANT ALL PRIVILEGES ON ALL SEQUENCES IN SCHEMA tenant_clinica1         TO clinica1_user;
-GRANT EXECUTE ON ALL FUNCTIONS IN SCHEMA tenant_clinica1           TO clinica1_user;
+GRANT ALL PRIVILEGES ON ALL TABLES IN SCHEMA tenant_clinica1 TO clinica1_user;
+GRANT ALL PRIVILEGES ON ALL SEQUENCES IN SCHEMA tenant_clinica1 TO clinica1_user;
+GRANT EXECUTE ON ALL FUNCTIONS IN SCHEMA tenant_clinica1 TO clinica1_user;
 
 SET search_path TO tenant_clinica1;
 
 CREATE TABLE admin_tenant (
-    admin_id SERIAL PRIMARY KEY,
+    admin_id BIGSERIAL PRIMARY KEY,
     nombre_usuario VARCHAR(50) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
     nombre VARCHAR(50),
@@ -22,7 +22,7 @@ CREATE TABLE admin_tenant (
 );
 
 CREATE TABLE profesional (
-    profesional_id SERIAL PRIMARY KEY,
+    profesional_id BIGSERIAL PRIMARY KEY,
     nombre_usuario VARCHAR(50) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
     nombre VARCHAR(50),
@@ -33,8 +33,8 @@ CREATE TABLE profesional (
 );
 
 CREATE TABLE paciente (
-    paciente_id SERIAL PRIMARY KEY,
-    globaluser_id INT,
+    paciente_id BIGSERIAL PRIMARY KEY,
+    globaluser_id BIGINT,
     nombre VARCHAR(50),
     apellido VARCHAR(50),
     telefono VARCHAR(15),
@@ -44,9 +44,9 @@ CREATE TABLE paciente (
 );
 
 CREATE TABLE documento_clinico (
-    doc_id SERIAL PRIMARY KEY,
-    paciente_id INT NOT NULL,
-    profesional_id INT NOT NULL,
+    doc_id BIGSERIAL PRIMARY KEY,
+    paciente_id BIGINT NOT NULL,
+    profesional_id BIGINT NOT NULL,
     id_externa_doc VARCHAR(120) UNIQUE,
     tipo VARCHAR(30),
     estado VARCHAR(30),
@@ -58,7 +58,7 @@ CREATE TABLE documento_clinico (
 );
 
 CREATE TABLE tenant_config (
-    config_id SERIAL PRIMARY KEY,
+    config_id BIGSERIAL PRIMARY KEY,
     nombre_visible VARCHAR(120),
     color_principal VARCHAR(40),
     logourl VARCHAR(300),
@@ -68,18 +68,18 @@ CREATE TABLE tenant_config (
 );
 
 CREATE TABLE central_endpoint (
-    endpoint_id SERIAL PRIMARY KEY,
+    endpoint_id BIGSERIAL PRIMARY KEY,
     url_base VARCHAR(300) NOT NULL,
     tipo_auth VARCHAR(40),
     hash_cliente VARCHAR(255)
 );
 
 CREATE TABLE audit_log (
-    audit_id SERIAL PRIMARY KEY,
+    audit_id BIGSERIAL PRIMARY KEY,
     tipo_actor VARCHAR(50),
-    actor_id INT NOT NULL,
+    actor_id BIGINT NOT NULL,
     accion VARCHAR(120),
-    recurso_id INT,
+    recurso_id BIGINT,
     resultado VARCHAR(80),
     ip VARCHAR(45),
     fecha_creacion TIMESTAMP
